Move files in the phone bill web original project to the "phonebillweb" package so that we can have unique package names in the original projects that will not confuse the aggregate javadoc build.
