+++ projects-parent/originals-parent/phonebill-web/pom.xml
@@ -160,7 +160,7 @@
         <configuration>
           <archive>
             <manifest>
-              <mainClass>edu.pdx.cs410J.original.Project4</mainClass>
+              <mainClass>edu.pdx.cs410J.phonebillweb.Project4</mainClass>
             </manifest>
           </archive>
         </configuration>
