+++ projects-parent/originals-parent/phonebill-web/src/it/java/edu/pdx/cs410J/phonebillweb/PhoneBillRestClientIT.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.phonebillweb;
 
 import edu.pdx.cs410J.web.HttpRequestHelper;
 import org.junit.FixMethodOrder;
