+++ projects-parent/originals-parent/phonebill-web/src/it/java/edu/pdx/cs410J/phonebillweb/Project4IT.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.phonebillweb;
 
 import edu.pdx.cs410J.InvokeMainTestCase;
 import edu.pdx.cs410J.web.HttpRequestHelper;
@@ -13,7 +13,7 @@ import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 
 /**
- * Tests the {@link Project4} class by invoking its main method with various arguments 
+ * Tests the {@link Project4} class by invoking its main method with various arguments
  */
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class Project4IT extends InvokeMainTestCase {
