+++ projects-parent/originals-parent/phonebill-web/src/main/java/edu/pdx/cs410J/phonebillweb/Messages.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.phonebillweb;
 
 /**
  * Class for formatting messages on the server side.  This is mainly to enable
