+++ projects-parent/originals-parent/phonebill-web/src/main/java/edu/pdx/cs410J/phonebillweb/PhoneBillServlet.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.phonebillweb;
 
 import com.google.common.annotations.VisibleForTesting;
