+++ projects-parent/originals-parent/phonebill-web/src/test/java/edu/pdx/cs410J/phonebillweb/PhoneBillServletTest.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.phonebillweb;
 
 import org.junit.Test;
 
@@ -10,9 +10,7 @@ import java.io.PrintWriter;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 /**
  * A unit test for the {@link PhoneBillServlet}.  It uses mockito to
